more fixes

+++ arcade/text.py
@@ -178,8 +178,8 @@ class Text:
     def __init__(
         self,
         text: str,
-        start_x: float,
-        start_y: float,
+        start_x: int,
+        start_y: int,
         color: RGBOrA255 = arcade.color.WHITE,
         font_size: float = 12,
         width: Optional[int] = 0,
@@ -193,7 +193,7 @@ class Text:
         rotation: float = 0,
         batch: Optional[pyglet.graphics.Batch] = None,
         group: Optional[pyglet.graphics.Group] = None,
-        start_z: float = 0
+        start_z: int = 0
     ):
         if align != "center" and align != "left" and align != "right":
             raise ValueError("The 'align' parameter must be equal to 'left', 'right', or 'center'.")
@@ -214,7 +214,7 @@ class Text:
             bold=bold,
             italic=italic,
             multiline=multiline,
-            rotation=rotation,
+            rotation=rotation, # type: ignore pending https://github.com/pyglet/pyglet/issues/843
             batch=batch,
             group=group
         )
